pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robot-visionaries"
version = "0.1.0"
authors = [
    { name = "Atharva Pawar", email = "atharva2185@gmail.com" },
]
description = "A computer vision library for controlling 4-DOF robotic arms using hand gestures."
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
keywords = ["robotics", "computer-vision", "mediapipe", "serial"]
dependencies = [
    "pyserial>=3.5",
    "opencv-python>=4.5.0",
    "mediapipe>=0.9.0",
]

[project.urls]
"Homepage" = "https://github.com/spy-7/robot-visionaries"
"Bug Tracker" = "https://github.com/spy-7/robot-visionaries/issues"

[tool.setuptools]
packages = ["robot_visionaries"]